+++ Libplanet/Net/Swarm.cs
@@ -1281,8 +1281,8 @@ namespace Libplanet.Net
                 }
 
                 txs = new HashSet<Transaction<T>>(
-                    txs.Where(tx => BlockChain.Policy.DoesTransactionFollowsPolicy(tx, BlockChain))
-                );
+                    txs.Where(
+                        tx => BlockChain.Policy.ValidateNextBlockTx(BlockChain, tx) is null));
 
                 foreach (Transaction<T> tx in txs)
                 {